Split Kimi runtime helpers
